light refactor
